requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via aurelio-sdk
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.12
    # via
    #   aurelio-sdk
    #   litellm
    #   semantic-router
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
aurelio-sdk==0.0.19
    # via semantic-router
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via litellm
colorama==0.4.6
    # via semantic-router
colorlog==6.9.0
    # via
    #   aurelio-sdk
    #   semantic-router
distro==1.9.0
    # via openai
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via huggingface-hub
h11==0.16.0
    # via httpcore
hf-xet==1.1.3
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   litellm
    #   openai
huggingface-hub==0.32.5
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
jinja2==3.1.6
    # via litellm
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via scikit-learn
jsonschema==4.24.0
    # via litellm
jsonschema-specifications==2025.4.1
    # via jsonschema
litellm==1.72.3
    # via semantic-router
markupsafe==3.0.2
    # via jinja2
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
numpy==2.3.0
    # via
    #   -r requirements.in
    #   pandas
    #   scikit-learn
    #   scipy
    #   semantic-router
openai==1.86.0
    # via
    #   -r requirements.in
    #   litellm
    #   semantic-router
packaging==25.0
    # via huggingface-hub
pandas==2.3.0
    # via -r requirements.in
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.5
    # via
    #   aurelio-sdk
    #   litellm
    #   openai
    #   semantic-router
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   aurelio-sdk
    #   litellm
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   semantic-router
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   semantic-router
    #   tiktoken
requests==2.32.4
    # via
    #   aurelio-sdk
    #   huggingface-hub
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via aurelio-sdk
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.7.0
    # via -r requirements.in
scipy==1.15.3
    # via scikit-learn
semantic-router==0.0.72
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.9.0
    # via
    #   litellm
    #   semantic-router
tokenizers==0.21.1
    # via litellm
tornado==6.5.1
    # via
    #   aurelio-sdk
    #   semantic-router
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   openai
typing-extensions==4.14.0
    # via
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==1.26.20
    # via
    #   requests
    #   semantic-router
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
